ADD Bonus Mark
At that moment, I had no idea we would add a standard library, and I wanted to have `fun`as a builtin.
I did know how to use mpc with a string, so I did it "by hand" (I created the arguments and passed them to the other builtin_* C functions). I found that exercise to be very interesting.