chore: update mpca_lang() usage with null terminated
In the current version of mpc parser, we can add NULL add the end of
arguments mpca_lang() to indicate that it's the end of argument when
trying to find parser and prevent segfault error.

This changes update the snippet for the web version and example that use
mpca_lang() by adding NULL at the end of mpca_lang() argument.